Lift the "default firmware filename" to ActiveDevice on the theory that the UI ought not to be touching the filesystem.
