+++ src/renderer/ActiveDevice.js
@@ -16,6 +16,8 @@
  */
 
 import Focus from "@api/focus";
+import path from "path";
+import { getStaticPath } from "@renderer/config";
 
 export function ActiveDevice() {
   this.port = undefined;
@@ -65,5 +67,18 @@ export function ActiveDevice() {
     } else {
       return false;
     }
+  };
+  this.defaultFirmwareFilename = () => {
+    const { vendor, product } = this.focus.focusDeviceDescriptor.info;
+    const firmwareType =
+      this.focus.focusDeviceDescriptor.info.firmwareType || "hex";
+    const cVendor = vendor.replace("/", ""),
+      cProduct = product.replace("/", "");
+    return path.join(
+      getStaticPath(),
+      cVendor,
+      cProduct,
+      "default." + firmwareType
+    );
   };
 }
